pagina_web/requirements.txt
Flask
numpy
requests
pandas
joblib
google-generativeai
scikit-learn